+++ src/SWP/Bundle/CoreBundle/Service/ArticlePublisher.php
@@ -109,7 +109,11 @@ final class ArticlePublisher implements ArticlePublisherInterface
                 $article->setRoute($destination->getRoute());
                 $article->setPublishedFBIA($destination->isFbia());
                 $this->eventDispatcher->dispatch(Events::SWP_VALIDATION, new GenericEvent($article));
-                $this->dispatchEvents($article, $package);
+                if ($destination->isPublished()) {
+                    $this->eventDispatcher->dispatch(ArticleEvents::PUBLISH, new ArticleEvent($article, null, ArticleEvents::PUBLISH));
+                }
+
+                $this->eventDispatcher->dispatch(ArticleEvents::PRE_CREATE, new ArticleEvent($article, $package, ArticleEvents::PRE_CREATE));
 
                 continue;
             }
@@ -126,7 +130,12 @@ final class ArticlePublisher implements ArticlePublisherInterface
             $article->setPublishedFBIA($destination->isFbia());
             $article->setArticleStatistics($articleStatistics);
             $this->articleRepository->persist($article);
-            $this->dispatchEvents($article, $package);
+
+            if ($destination->isPublished()) {
+                $this->eventDispatcher->dispatch(ArticleEvents::PUBLISH, new ArticleEvent($article, null, ArticleEvents::PUBLISH));
+            }
+
+            $this->eventDispatcher->dispatch(ArticleEvents::PRE_CREATE, new ArticleEvent($article, $package, ArticleEvents::PRE_CREATE));
         }
 
         $this->articleRepository->flush();
@@ -144,15 +153,5 @@ final class ArticlePublisher implements ArticlePublisherInterface
             'tenantCode' => $tenantCode,
             'code' => $code,
         ]);
-    }
-
-    /**
-     * @param ArticleInterface $article
-     * @param PackageInterface $package
-     */
-    private function dispatchEvents(ArticleInterface $article, PackageInterface $package)
-    {
-        $this->eventDispatcher->dispatch(ArticleEvents::PUBLISH, new ArticleEvent($article, null, ArticleEvents::PUBLISH));
-        $this->eventDispatcher->dispatch(ArticleEvents::PRE_CREATE, new ArticleEvent($article, $package, ArticleEvents::PRE_CREATE));
     }
 }