+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog references the relevant changes (bug and security fixes) done in 
 To get the diff for a specific change, go to https://github.com/superdesk/web-publisher/commit/XXX where XXX is the change hash
 
 * 1.1.x
+ * feature [#454] Override existing rules by publish destinations
  * improvement [#455] Add slug to article author - allow to load authors by slug
  * feature [#455] Allow to create custom routes (variable prefix and requirements)
  * feature [#451] Allowed to evaluate rules that match given package